print-locks.pl: initial commit
This script helps to find possible deadlocks, due to false locking.
It has a hard-coded list of locks (spinlocks and mutexes) that are
present in the source code. Via command line a source file is provided
and the script finds any occurences of the given locks and prints
those functions that make use of them and call other functions, while
holding the lock.
This helps to find situations where the locking order is being violated.

Signed-off-by: Christian Gromm <christian.gromm@microchip.com>